added more tests, refactored, and moved some useful stuff into demo_statkraft_stuff.py
