+++ shyft/repository/service/gis_region_model_repository.py
@@ -552,92 +552,4 @@ class GisRegionModelRepository(RegionModelRepository):
         return region_model
 
 
-def add_plot_polygons(ax, polygons, color):
-    ps = []
-    if not isinstance(polygons, list):
-        polygons = [polygons]
-    for polygon in polygons:
-        if isinstance(polygon, MultiPolygon):
-            ps.extend([p for p in polygon])
-        else:
-            ps.append(polygon)
-    patches = imap(PolygonPatch, [p for p in ps])
-    ax.add_collection(PatchCollection(patches, facecolors=(color,), linewidths=0.1, alpha=0.3))
 
-def run_cell_example(catchment_type,identifier,x0, y0, dx, dy, nx, ny, catch_indicies,epsg_id):
-    grid_spec=GridSpecification(x0, y0, dx, dy, nx, ny)
-    cf = CellDataFetcher(catchment_type,identifier,grid_spec, id_list=catch_indicies,epsg_id=epsg_id)
-    print( "Start fetching data")
-    cf.fetch()
-    print ("Done, now preparing plot")
-    # Plot the extracted data
-    fig, ax = plt.subplots(1)
-    color_map = {"forest": 'g', "lake": 'b', "glacier": 'r', "cell": "0.75", "reservoir": "purple"}
-
-    extent = grid_spec.geometry[0], grid_spec.geometry[2], grid_spec.geometry[1], grid_spec.geometry[3]
-    ax.imshow(cf.elevation_raster, origin='upper', extent=extent, cmap=cm.gray)
-
-    for catchment_cells in cf.cell_data.itervalues():
-        add_plot_polygons(ax, [cell["cell"] for cell in catchment_cells], color=color_map["cell"])
-    for catchment_land_type in cf.catchment_land_types.itervalues():
-        for k,v in catchment_land_type.iteritems():
-            add_plot_polygons(ax, v, color=color_map[k])
-
-    geometry = grid_spec.geometry
-    ax.set_xlim(geometry[0], geometry[2])
-    ax.set_ylim(geometry[1], geometry[3])
-    plt.show()
-
-def nea_nidelv_example(epsg_id):
-    x0 = 270000.0
-    y0 = 6960000.0
-    dx = 1000
-    dy = 1000
-    nx = 105
-    ny = 75
-    # test fetching for regulated catchments using CATCH_ID
-    id_list=[1228,1308,1394,1443,1726,1867,1996,2041,2129,2195,2198,2277,2402,2446,2465,2545,2640,2718,3002,3536,3630,1000010,1000011]
-    run_cell_example('regulated','CATCH_ID',x0, y0, dx, dy, nx, ny, id_list,epsg_id=epsg_id)
-    # test fetching for regulated catchments using POWER_PLANT_ID
-    ##id_list=[38, 87, 115, 188, 259, 291, 292, 295, 389, 465, 496, 516, 551, 780]
-    #id_list=[38,115,137,188,291,292,389,465,496,551,780,1371,1436]   
-    #run_cell_example('regulated','POWER_PLANT_ID',x0, y0, dx, dy, nx, ny, id_list,epsg_id=epsg_id)
-    # test fetching for unregulated catchments using FELTNR
-    #id_list=[1691,1686]
-    #run_cell_example2('unregulated','FELTNR',x0, y0, dx, dy, nx, ny, id_list,epsg_id=epsg_id)
-    
-def vinjevatn_example(epsg_id):
-    x0 = 73000.0
-    y0 = 6613000.0
-    dx = 1000
-    dy = 1000
-    nx = 40
-    ny = 35
-    # test fetching for regulated catchments using CATCH_ID
-    #id_list=[2668,2936,2937,2938,2939,2940,2941,2942]
-    #run_cell_example('regulated','CATCH_ID',x0, y0, dx, dy, nx, ny, id_list,epsg_id=epsg_id)
-    # test fetching for regulated catchments using POWER_PLANT_ID
-    id_list=[446]
-    run_cell_example('regulated','POWER_PLANT_ID',x0, y0, dx, dy, nx, ny, id_list,epsg_id=epsg_id)
-    # test fetching for unregulated catchments using FELTNR
-    #id_list=[645,702]
-    #run_cell_example2('unregulated','FELTNR',x0, y0, dx, dy, nx, ny, id_list,epsg_id=epsg_id)
-    
-    
-def test_example():
-    x0 = 350000.0
-    y0 = 6751000.0
-    dx = 1000
-    dy = 1000
-    nx = 32
-    ny = 24
-    #id_list = [163,287,332]
-    #run_cell_example2('regulated','POWER_PLANT_ID',x0, y0, dx, dy, nx, ny, id_list)
-    id_list = [1225,1226]
-    run_cell_example('unregulated','FELTNR',x0, y0, dx, dy, nx, ny, id_list)
-
-if __name__ == "__main__":
-    #vinjevatn_example(32633)
-    nea_nidelv_example(32633)
-    #test_example()
-    #dtm_example()
